EWahaAiApi/pyproject.toml
[tool.poetry]
name = "verysmartapi"
version = "0.1.0"
description = ""
authors = ["Bartosz Michałowski <bartoszm.dev@tutanota.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.3"
fastapi = "^0.63.0"
SQLAlchemy = "^1.3.22"
azure-cognitiveservices-vision-computervision = "^0.7.0"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
